# MAINT: Make constant and edge mode padding faster.