Kept only as reference, while https://github.com/numpy/numpy/pull/11358 is being developed.

This is a start at addressing #11126

Padding was causing the array to be copied `2*arr.ndim` times which can be huge for large arrays.

I did not comment my helper functions. 
They are 1-2 line long.
If numpy requires those to be commented then please let me know.

Modes finished:
  * [x] constant
  * [x] edge
  * [x] linear_ramp
  * [x] maximum
  * [x] mean
  * [x] median
  * [x] minimum
  * [ ] reflect
  * [ ] symmetric
  * [ ] wrap

I'm glad to pass on the torch.

Before
```
================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========
--                                                                        mode                                                      
------------------ -----------------------------------------------------------------------------------------------------------------
    shape         constant     edge     linear_ramp   maximum      mean      median    minimum    reflect    symmetric     wrap   
================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========
   (64, 64)       625±0μs    747±0μs      1.32±0ms    1.03±0ms   1.22±0ms   2.51±0ms   1.41±0ms   506±0μs     617±0μs    618±0μs  
  (256, 256)      2.57±0ms   3.02±0ms     4.44±0ms    4.56±0ms   4.50±0ms   9.95±0ms   4.61±0ms   2.69±0ms    2.71±0ms   2.59±0ms 
  (512, 512)      8.38±0ms   6.59±0ms     7.64±0ms    9.02±0ms   7.15±0ms   19.3±0ms   11.4±0ms   8.07±0ms    6.77±0ms   8.64±0ms 
 (10, 10, 10)     1.31±0ms   1.02±0ms     2.56±0ms    883±0μs    2.09±0ms   3.74±0ms   1.81±0ms   1.28±0ms    892±0μs    853±0μs  
 (10, 64, 64)     5.88±0ms   8.69±0ms     12.2±0ms    9.89±0ms   6.38±0ms   13.3±0ms   10.8±0ms   5.91±0ms    5.16±0ms   4.39±0ms 
(50, 1024, 1024)   1.88±0s    1.85±0s      2.20±0s     2.23±0s    2.15±0s    5.63±0s    2.24±0s    1.11±0s     1.10±0s    1.11±0s  
================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========
```

After (only constant and edge were implemented)
```


================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========
--                                                                        mode                                                      
------------------ -----------------------------------------------------------------------------------------------------------------
    shape         constant     edge     linear_ramp   maximum      mean      median    minimum    reflect    symmetric     wrap   
================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========
   (64, 64)       840±0μs    825±0μs      916±0μs     962±0μs    858±0μs    2.10±0ms   667±0μs    595±0μs     655±0μs    634±0μs  
  (256, 256)      1.94±0ms   1.91±0ms     4.45±0ms    4.57±0ms   4.41±0ms   5.65±0ms   3.16±0ms   2.46±0ms    1.11±0ms   2.57±0ms 
  (512, 512)      3.00±0ms   4.77±0ms     10.4±0ms    10.4±0ms   9.34±0ms   21.4±0ms   10.6±0ms   6.99±0ms    6.77±0ms   4.52±0ms 
 (10, 10, 10)     1.10±0ms   1.14±0ms     2.57±0ms    1.83±0ms   1.49±0ms   3.84±0ms   1.83±0ms   1.30±0ms    938±0μs    605±0μs  
 (10, 64, 64)     2.66±0ms   3.33±0ms     10.7±0ms    8.38±0ms   7.58±0ms   14.5±0ms   8.83±0ms   5.67±0ms    4.82±0ms   5.01±0ms 
(50, 1024, 1024)   411±0ms    419±0ms      2.20±0s     2.23±0s    2.15±0s    5.69±0s    2.25±0s    1.12±0s     1.11±0s    1.08±0s  
================== ========== ========== ============= ========== ========== ========== ========== ========== =========== ==========

```

cc @jakirkham