@eric-wieser 2018-09-15
This seems to overlap heavily with the goals of #11358