@hmaarrfk 2018-09-15
Man. Totally didn’t see that PR. Ill have to compare as to how it fairs with the performance of this one.