@eric-wieser 2018-09-15
I think I'd like to see a version of this built on top of #11955 - the one-copy approach is neat, but I think everything would become less convoluted if we could separate calculating the padded values from inserting them