@hmaarrfk 2018-09-15
Hey Eric. 

Thanks for having a look ar my PR. It is a little late for me but ill try to take a look at your PR tomorrow.

Mark